usak-natural-gas-clustering/requirements.txt
pandas>=1.5
numpy>=1.23
scikit-learn>=1.1
scipy>=1.9
matplotlib>=3.6
openpyxl>=3.0
folium>=0.14
streamlit>=1.20
streamlit-option-menu>=0.3.6
Pillow>=9.4
